pyproject.toml
[project]
name = "cv"
version = "6.0.0"
description = "Build my resume page"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livereload>=2.7.1",
    "markdown>=3.7",
    "py-buzz>=5.0.2",
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
    "typer>=0.13.0",
]

[tool.uv]
package = true


[dependency-groups]
dev = [
    "ipython>=9.0.1",
    "pyclean>=3.1.0",
]

[project.scripts]
cv = "cv.main:cli"

[tool.hatch.build.targets.wheel]
packages = [
    "src/cv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
